requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
fpdf==1.7.2
idna==3.2
img2pdf==0.4.1
lxml==4.6.3
numpy==1.21.1
pandas==1.3.1
pikepdf==2.16.1
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.0
urllib3==1.26.6
